Changed every float to double

+++ CPorting/Utilities.h
@@ -4,22 +4,22 @@
 #include <stdio.h>
 
 //Finds minimum, maximum and range of a float vector  
-void findMinMax(float *map, int n, float *min, float *max, float * range);
+void findMinMax(double *map, int n, double *min, double *max, double * range);
 
 //performs confrontation between and b given a single char operation
-int confront(float a, float b, char operation);
+int confront(double a, double b, char operation);
 
 //Gets a binary vector in matlab style, given a certain operation
-int * getBinaryVector(float * map, int n, int (*confront)(float, float), float value, int * numOfElementsMatching);
+int * getBinaryVector(double * map, int n, int (*confront)(double, double), double value, int * numOfElementsMatching);
 
 //Given a binary vector and a real vector, fills a new vector in the indices where binary == 1
-float * fromBinaryToRealVector(float * map, int n, int * binaryVector);
+double * fromBinaryToRealVector(double * map, int n, int * binaryVector);
 
 //creates a n-element vector of float filled with 0
-float * fill0(int n);
+double * fill0(int n);
 
 //Applies a single char operation on every element of the vector
-void apply_function(float * vector, int n, float (* operation) (float a, float b), float argument);
+void apply_function(double * vector, int n, double (* operation) (double a, double b), double argument);
 
 //converts 3D to 1D index
 int linearIndexFromCoordinate(int x, int y, int z, int max_x, int max_y);
@@ -27,19 +27,19 @@ int linearIndexFromCoordinate(int x, int y, int z, int max_x, int max_y);
 void coordinatesFromLinearIndex(int index, int max_x, int max_y, int * x, int * y, int * z);
 
 //copies an int vector into a float vector
-float * copyAndConvertIntVector(int * vector, int n);
+double * copyAndConvertIntVector(int * vector, int n);
 #endif //UTILITIES_H
 
-int lessThan(float a, float b);
+int lessThan(double a, double b);
 
-int equalTo(float a, float b);
+int equalTo(double a, double b);
 
-int moreThan(float a, float b);
+int moreThan(double a, double b);
 
-float multiply(float a, float b);
+double multiply(double a, double b);
 
-float elevate(float a, float b);
+double elevate(double a, double b);
 
-void abs_vector(float * vector, int n);
+void abs_vector(double * vector, int n);
 
-float * readMatFromFile(FILE * fp, int * size, int * x, int * y, int * z);
+double * readMatFromFile(FILE * fp, int * size, int * x, int * y, int * z);
